+++ app/Jobs/Issues/SendEmailToFollowersAboutReminder.php
@@ -41,31 +41,31 @@ class SendEmailToFollowersAboutReminder implements ShouldQueue
      */
     public function handle()
     {
-        Log::info('Job: SendEmailToFollowersAboutReminder. ' . $this->issue->ticketNumber);
+        Log::channel('templog')->debug('Job: SendEmailToFollowersAboutReminder. ' . $this->issue->ticketNumber);
         if (!is_null($this->issue->timeClosed)) {
-            Log::info(' - Avslutat. Inget mail skickat');
+            Log::channel('templog')->debug(' - Avslutat. Inget mail skickat');
             return;
         }
-        Log::info(' - typeOfReminder=' . $this->typeOfReminder);
+        Log::channel('templog')->debug(' - typeOfReminder=' . $this->typeOfReminder);
 
         switch ($this->typeOfReminder) {
             case 'paused':
                 if (!$this->issue->paused) {
-                    Log::info('   ' . $this->typeOfReminder . ' inte aktuell längre. Inget mail skickat');
+                    Log::channel('templog')->debug('   ' . $this->typeOfReminder . ' inte aktuell längre. Inget mail skickat');
                     return null;
                 }
                 $delayDays = setting('days_reminder_paused_issue');
                 break;
             case 'waitingForInternal':
                 if (!$this->issue->waitingForInternal) {
-                    Log::info('   ' . $this->typeOfReminder . ' inte aktuell längre. Inget mail skickat');
+                    Log::channel('templog')->debug('   ' . $this->typeOfReminder . ' inte aktuell längre. Inget mail skickat');
                     return null;
                 }
                 $delayDays = setting('days_reminder_waiting_for_internal');
                 break;
             case 'waitingForCustomer':
                 if (!$this->issue->waitingForCustomer) {
-                    Log::info('   ' . $this->typeOfReminder . ' inte aktuell längre. Inget mail skickat');
+                    Log::channel('templog')->debug('   ' . $this->typeOfReminder . ' inte aktuell längre. Inget mail skickat');
                     return null;
                 }
                 $delayDays = setting('days_reminder_waiting_for_external');
@@ -73,38 +73,38 @@ class SendEmailToFollowersAboutReminder implements ShouldQueue
 
             default:
                 if ($this->issue->paused) {
-                    Log::info('   Ärende markerat som pausat. Inget mail skickat');
+                    Log::channel('templog')->debug('   Ärende markerat som pausat. Inget mail skickat');
                     return null;
                 }
                 if ($this->issue->waitingForInternal) {
-                    Log::info('   Ärende markerat som väntar på kollega. Inget mail skickat');
+                    Log::channel('templog')->debug('   Ärende markerat som väntar på kollega. Inget mail skickat');
                     return null;
                 }
                 if ($this->issue->waitingForCustomer) {
-                    Log::info('   Ärende markerat som väntar på kund. Inget mail skickat');
+                    Log::channel('templog')->debug('   Ärende markerat som väntar på kund. Inget mail skickat');
                     return null;
                 }
                 $delayDays = setting('days_reminder_waiting_for_comment');
                 $delayDateTime = nextWorkingDateTime(workDaysToMinutes($delayDays));
                 if (cache($this->issue->ticketNumber . 'Cold')) {
-                    Log::info('   Cache-key: ' . $this->issue->ticketNumber . 'Cold exist. Inget mail skickat');
+                    Log::channel('templog')->debug('   Cache-key: ' . $this->issue->ticketNumber . 'Cold exist. Inget mail skickat');
                     return null;
                 }
                 if (is_null($this->issue->latestComment)) {
-                    Log::info('   Finns ingen kommentar');
+                    Log::channel('templog')->debug('   Finns ingen kommentar');
                 } else {
                     if (nextWorkingDateTime(workDaysToMinutes($delayDays) - 1, $this->issue->latestComment->updated_at) > nextWorkingDateTime()) {
-                        Log::info('   Senaste kommentar: ' . $this->issue->latestComment->updated_at);
-                        Log::info('   Mail förväntas skickas: ' . nextWorkingDateTime(workDaysToMinutes($delayDays), $this->issue->latestComment->updated_at));
-                        Log::info('   Inget mail skickat');
+                        Log::channel('templog')->debug('   Senaste kommentar: ' . $this->issue->latestComment->updated_at);
+                        Log::channel('templog')->debug('   Mail förväntas skickas: ' . nextWorkingDateTime(workDaysToMinutes($delayDays), $this->issue->latestComment->updated_at));
+                        Log::channel('templog')->debug('   Inget mail skickat');
                         return null;
                     }
                 }
                 cache([$this->issue->ticketNumber . 'Cold' => true], $delayDateTime);
-                Log::info('   Cache-key updated: ' . $this->issue->ticketNumber . 'Cold' . '. Expires: ' . $delayDateTime);
+                Log::channel('templog')->debug('   Cache-key updated: ' . $this->issue->ticketNumber . 'Cold' . '. Expires: ' . $delayDateTime);
                 break;
         }
         Mail::to($this->email)->send(new MailToFollowersAboutReminder($this->issue, $this->typeOfReminder));
-        Log::info('   MailToFollowersAboutReminder skickas: ' . $this->email);
+        Log::channel('templog')->debug('   MailToFollowersAboutReminder skickas: ' . $this->email);
     }
 }