+++ app/Jobs/Issues/SendEmailToFollowersAboutUpdate.php
@@ -43,6 +43,6 @@ class SendEmailToFollowersAboutUpdate implements ShouldQueue
     public function handle()
     {
        Mail::to($this->email)->send(new MailToFollowersAboutUpdate($this->issue, $this->type));
-       Log::info('   MailToFollowersAboutUpdate skickas: '.$this->email);
+       Log::channel('templog')->debug('   MailToFollowersAboutUpdate skickas: '.$this->email);
     }
 }